requirements.txt
absl-py==1.3.0
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bidict==0.21.4
Bottleneck==1.3.2
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.2.1
certifi==2023.5.7
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.4
conda==4.11.0
conda-package-handling==1.7.3
cryptography==36.0.0
cycler==0.11.0
datasets==1.18.4
decorator==5.1.1
dgl==1.1.0+cu113
dictdiffer==0.9.0
diffusers==0.12.1
dill==0.3.4
docker-pycreds==0.4.0
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.6.0
Flask==2.0.3
Flask-Compress==1.11
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.2.0
gdown==4.6.0
gensim==4.1.2
getmac==0.8.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.16.0
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
grpcio==1.51.1
hiplot==0.1.32
huggingface-hub==0.12.0
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipython==8.1.1
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.17.3
jupyter_core==5.0.0
kiwisolver==1.3.2
lightning-utilities==0.9.0
littleutils==0.2.2
llvmlite==0.38.0
lmdb==1.4.0
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.12.2
nbformat==5.7.1
networkx==2.6.3
numba==0.55.1
numexpr==2.8.1
numpy==1.22.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-htop==1.0.5
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ogb==1.3.5
openpyxl==3.0.9
outdated==0.2.1
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-louvain==0.15
pytz==2021.3
PyYAML==6.0
rdkit==2023.3.3
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
ruamel-yaml-conda==0.15.100
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentencepiece==0.1.97
sentry-sdk==1.5.6
setproctitle==1.2.2
setuptools==58.0.4
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
sympy==1.12
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
threadpoolctl==2.2.0
timeloop==1.0.2
tokenizers==0.11.6
#torch==1.10.2
torch-cluster==1.5.9
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchaudio==0.10.2
torchmetrics==1.2.0
torchvision==0.11.3
tqdm==4.62.3
traitlets==5.1.1
transformers==4.25.1
tsc-auto==0.52
typing_extensions==4.8.0
urllib3==1.26.8
wandb==0.12.11
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.1
xxhash==3.0.0
yacs==0.1.6
yarl==1.7.2
yaspin==2.1.0
zipp==3.11.0
